Add files via upload
Housing prices closely relate to people’s livelihood and the economic trend of a nation. Having a general idea of the trend of the housing price level is crucial nowadays no matter for individuals or the government. Thus, realizing accurate housing price prediction is essential and it is helpful for buyers or officials to make wiser decisions in terms of investing or policymaking. In this paper, a hybrid model is designed to predict the Perth housing prices with more comprehensive features related to location and time series. The model includes a price model predicting the time series trend and an adjusted model to predict housing prices with regression machine learning methods. For the price model, we implement Gated Recurrent Units (GRU) and Long Short-Term Memory (LSTM) neural network methods. The corresponding predicted monthly mean price will be added to the adjusted model as a feature. For the adjusted model, we apply four classic regression models to do the comparative analysis. By evaluating the R-square and the mean average percentage error, we obtain a prediction model with over 80% accuracy, which may contribute to the study in the real estate field.